Kill markdocs service only when it has already started

+++ docs-preview/src/server.ts
@@ -71,7 +71,10 @@ export class MarkdocsServer {
     }
 
     public stopMarkdocsServer() {
-        this.spawnProcess.kill();
+        if (this.hasAlreadyStartAsync())
+        {
+            this.spawnProcess.kill();
+        }
     }
 
     private async ensureMarkdocsServerWorkAsync(): Promise<void> {
